Polish OpenAI extraction validation fixes
Agent-Logs-Url: https://github.com/NameHaibinZhang/opentelemetry-ebpf-instrumentation/sessions/4c4e4c96-a04c-407e-bec7-8145acabeaca

+++ pkg/ebpf/common/http/responses.go
@@ -110,7 +110,7 @@ func getRequestBody(req *http.Request) ([]byte, error) {
 		return body, err
 	}
 
-	if int64(len(body)) > maxCapturedRequestBodyBytes {
+	if len(body) > int(maxCapturedRequestBodyBytes) {
 		body = body[:maxCapturedRequestBodyBytes]
 		req.Body = io.NopCloser(bytes.NewBuffer(body))
 		return body, errRequestBodyTooLarge
